# State Machine: Optimize negative lookups in the account/transfer creation prefetch flow