### Problem

While creating a new account/transfer, the state machine ascertains the existence of the account/transfer ID in the LSM tree. If the ID doesn't exist in the tree, the account/transfer can be safely created. Otherwise, the account/transfer creation must be rejected.

 To this end,  [prefetch_create_accounts](https://github.com/tigerbeetledb/tigerbeetle/blob/f7e64c970a6ff38987bc64dce92ce519ee93e64e/src/state_machine.zig#L407) & [prefetch_create_transfers](https://github.com/tigerbeetledb/tigerbeetle/blob/f7e64c970a6ff38987bc64dce92ce519ee93e64e/src/state_machine.zig#L439) invoke [prefetch_enqueue](https://github.com/tigerbeetledb/tigerbeetle/blob/f7e64c970a6ff38987bc64dce92ce519ee93e64e/src/lsm/groove.zig#L617), which involves searching for the ID in all on-disk levels of the LSM tree (if not found in the in-memory (im)mutable tables), possibly performing I/O to read index, filter & data blocks (if not found in the cache).

This is expensive, considering duplicate IDs are a rarity and would likely only be encountered if:

* The client retries a batch of requests for which it didn't receive an acknowledge
* The client generates duplicate IDs within the same batch

### Solution

[prefetch_enqueue_negative](https://github.com/tigerbeetledb/tigerbeetle/blob/f7e64c970a6ff38987bc64dce92ce519ee93e64e/src/lsm/groove.zig#L577) makes a best-effort attempt to ascertain the existence of an ID in the IdTree without performing any I/O. If ascertaining the existence of the ID becomes intractable at any point (this point is expanded on below), it reverts to the legacy flow.

_**This is achieved by maintaining key ranges for the Tree & all levels in the LSM tree.**_ 

Note that we don't maintain a key range for the mutable table. Maintaining a key range for the mutable table would entail maintaining the minimum & maximum key across a hash map and a SetAssociativeCache. While it is feasible to track the key range in the hash map, we would have to keep a track of evictions from the cache to maintain a key range for it, which is non trivial. This would become easier to track when @cb22's changes come in since they move the cache up to the Groove layer!

This is a high-level writeup of how the prefetch negative lookup flow looks after these changes get in:

* **_If the account/transfer ID does not exist within the tree's key range_** → the account/transfer creation request can simply be processed without incurring any I/O cost 🚀
* **_If the account/transfer ID does exist in the tree's key range, we handle the following cases:_**
  * _ID doesn't exist in the (im)mutable tables_ 
    * ID doesn't exist exist within any of the key ranges of the LSM levels → The account/transfer creation request can be processed without incurring any I/O cost 🚀
    * ID exists within the key range of some LSM levels → Attempt to synchronously fetch the index, data & filter blocks of the tables that could contain the key from the cache
      *  If any of the blocks required to ascertain the existence of a key within a table are not found in the cache, revert to the legacy flow
      *  If all the blocks required to ascertain the existence of a key within a table are found in the cache, and the ID doesn't exist in any of them → The account/transfer creation request can be processed without incurring any I/O cost 🚀
  * _ID exists in the (im)mutable tables_ → Attempt to prefetch it from the (im)mutable tables. Note that this step the same as it was in the legacy flow, since it attempts to prefetch the ID from memory before checking the on-disk LSM tree levels.