@chaitanyabhandari 2023-07-01
### With negative-lookup optimization

```
~/Desktop/tigerbeetle_summer23/tigerbeetle cbb-prefetch-negative-lookups* ⇡ 3m 40s ❯ sudo scripts/benchmark.sh                    11:18:22 PM

Formatting replica 0...
Starting replica 0...

Benchmarking...
info: Benchmark running against { 127.0.0.1:64324 }
info(message_bus): connected to replica 0
1223 batches in 99.51 s
load offered = 1000000 tx/s
load accepted = 100493 tx/s
batch latency p00 = 0 ms
batch latency p10 = 18 ms
batch latency p20 = 19 ms
batch latency p30 = 19 ms
batch latency p40 = 20 ms
batch latency p50 = 20 ms
batch latency p60 = 21 ms
batch latency p70 = 21 ms
batch latency p80 = 22 ms
batch latency p90 = 24 ms
batch latency p100 = 4253 ms
transfer latency p00 = 0 ms
transfer latency p10 = 4010 ms
transfer latency p20 = 7918 ms
transfer latency p30 = 14226 ms
transfer latency p40 = 22564 ms
transfer latency p50 = 33246 ms
transfer latency p60 = 44142 ms
transfer latency p70 = 51598 ms
transfer latency p80 = 63907 ms
transfer latency p90 = 76184 ms
transfer latency p100 = 89522 ms

Stopping replica 0...
```

### Without negative-lookup optimization

```
~/Desktop/tigerbeetle_summer23/tigerbeetle main* ❯ sudo scripts/benchmark.sh

Formatting replica 0...
Starting replica 0...

Benchmarking...
info: Benchmark running against { 127.0.0.1:64513 }
info(message_bus): connected to replica 0
1223 batches in 107.11 s
load offered = 1000000 tx/s
load accepted = 93362 tx/s
batch latency p00 = 0 ms
batch latency p10 = 22 ms
batch latency p20 = 23 ms
batch latency p30 = 24 ms
batch latency p40 = 25 ms
batch latency p50 = 26 ms
batch latency p60 = 27 ms
batch latency p70 = 28 ms
batch latency p80 = 29 ms
batch latency p90 = 30 ms
batch latency p100 = 4264 ms
transfer latency p00 = 0 ms
transfer latency p10 = 4610 ms
transfer latency p20 = 9137 ms
transfer latency p30 = 16160 ms
transfer latency p40 = 25441 ms
transfer latency p50 = 36809 ms
transfer latency p60 = 48459 ms
transfer latency p70 = 56894 ms
transfer latency p80 = 69369 ms
transfer latency p90 = 82695 ms
transfer latency p100 = 97122 ms

Stopping replica 0...
```