@chaitanyabhandari 2023-07-05
@dj thank you for bringing up https://github.com/tigerbeetle/tigerbeetle/pull/944#issuecomment-1621813796!

We did discuss this over call today well! w were thinking about how much performance gain we can attribute to using [read_from_cache](https://github.com/tigerbeetle/tigerbeetle/blob/c33789942e5f3f054e6391cafb5589268f30acbd/src/lsm/grid.zig#L496) instead of [read_block](https://github.com/tigerbeetle/tigerbeetle/blob/c33789942e5f3f054e6391cafb5589268f30acbd/src/lsm/grid.zig#L513). The purpose that it was supposed to solve (apart from helping random IDs) that it does so while avoiding some [queueing](https://github.com/tigerbeetle/tigerbeetle/blob/c33789942e5f3f054e6391cafb5589268f30acbd/src/lsm/grid.zig#L537-L550) and [callback resolution](https://github.com/tigerbeetle/tigerbeetle/blob/c33789942e5f3f054e6391cafb5589268f30acbd/src/lsm/grid.zig#L728-L734) overheads. I agree, it is a lot of added dimensionality and it makes sense measuring the performance without the added complexity, to figure out how much it's really contributing to performance.

I am planning on commenting out the [synchronous lookup](https://github.com/tigerbeetle/tigerbeetle/blob/c33789942e5f3f054e6391cafb5589268f30acbd/src/lsm/tree.zig#L317-L353) logic and measuring the performance when we only use the _Tree key range_ & _ManifestLevel key ranges_ for negative lookup.

cc: @batiati @jorangreef