@sentientwaffle 2023-07-05
GitHub inexplicably won't let me reply to this comment directly, sorry..

> For optimization, it may help to make levels a bitmask instead. Each bit can be set with mask |= boolToInt(exists(key)) << i. If key_may_exists is branchless, it could be vectorized depending on compare_keys.

This is an excellent idea for a reason completely independent of performance:

Right now, `LookupIterator.level` is misleading because it is being used as two logically different things depending on the caller: either an array index or a level index.

If instead of a `BoundedArray`, we use a bitset (with set bits to mark which levels should be covered) then it can default to `initFull()`. Then there's no added dimensionality for the second code path, just use `.isSet(level)` in `LevelIterator` to filter which levels are iterated.