@jorangreef 2023-07-05
> what do both sets of benchmarks look like without Tree.key_may_exist's manifest check, just the tree check?

@sentientwaffle great idea to check this.

> In particular it adds an entire duplicate (sync) reimplementation of the lookup path, which is a lot of added dimensionality.

I realized today concurrently that we can in fact use only this new sync lookup path, to get rid of the machinery of next tick for the prefetch path, provided we have a way to do direct grid reads, i.e. telling the grid to go direct to disk, since we know after prefetch that it's not already in cache.

Knowing whether it's in cache also applies as an optimization for positive lookups, enabling the async prefetch path to be bypassed.

> Especially since (I assume) it mainly helps fully-random ids, which we are not going to be recommending anyway.

We may need to support random workloads nevertheless. The other use case for this also is where the tree shape is impacted by an outlying id, for some reason, but then most ids are still clustered and partitioned, so that negative lookups for the most part can be optimized.