@chaitanyabhandari 2023-07-05
@sentientwaffle, somehow I'm not able to reply to https://github.com/tigerbeetle/tigerbeetle/pull/944#discussion_r1253180171 as well. I agree, using a `StaticBitSet` is an elegant solution. I had started thinking about some ways to not increase dimensionality in `LookupIterator` but the thought somehow got away from me. :)

Thanks @kprotty for the great input on that front!