@chaitanyabhandari 2023-07-11
## Microbenchmarking the prefetch flow

These microbenchmarks were conducted after augmenting the positive lookup path with synchronous cache reads (along with the negative lookup path, which was the initial change proposed in this PR).

### Random IDs

#### Without negative & positive lookup optimization

```C
latency p00 = 0 ms
latency p10 = 4 ms
latency p20 = 10 ms
latency p30 = 10 ms
latency p40 = 13 ms
latency p50 = 14 ms
latency p60 = 14 ms
latency p70 = 14 ms
latency p80 = 14 ms
latency p90 = 15 ms
latency p100 = 16 ms
```

#### With negative & positive lookup optimization

```C
latency p00 = 0 ms
latency p10 = 3 ms
latency p20 = 7 ms
latency p30 = 8 ms
latency p40 = 9 ms
latency p50 = 10 ms
latency p60 = 10 ms
latency p70 = 10 ms
latency p80 = 10 ms
latency p90 = 10 ms
latency p100 = 12 ms
```
### Monotonic IDs

#### Without negative & positive lookup optimization
```C
latency p00 = 0 ms
latency p10 = 4 ms
latency p20 = 7 ms
latency p30 = 7 ms
latency p40 = 7 ms
latency p50 = 7 ms
latency p60 = 7 ms
latency p70 = 7 ms
latency p80 = 7 ms
latency p90 = 7 ms
latency p100 = 14 ms
```

#### With negative & positive lookup optimization

```C
latency p00 = 0 ms
latency p10 = 3 ms
latency p20 = 5 ms
latency p30 = 5 ms
latency p40 = 5 ms
latency p50 = 5 ms
latency p60 = 5 ms
latency p70 = 5 ms
latency p80 = 5 ms
latency p90 = 5 ms
latency p100 = 13 ms
```

cc: @jorangreef @sentientwaffle @cb22