@sentientwaffle 2023-07-11
Can you squash all your commits? The merge commits mess up my editor's diff :stuck_out_tongue: 
(We generally try to rebase to pull changes from `main` into a feature branch, btw.)